Cleanup
I forgot to edit the function names of the new NMT view in the last commit so I'm fixing these in this.
I'm not very sure if there is more to do.
I also found a deprecated function (UI_NMT__init_nmtState) but I will leave it there in case we need to do anything with it